feat: Update bot_state table schema and enhance API error handling
- Changed the data type of `account_id` in the `bot_state` table from INTEGER to BIGINT to align with the accounts table schema.
- Improved error handling across various API endpoints by implementing a centralized logging function to capture and report database errors more effectively.
- Refactored API handlers to utilize the new error logging mechanism, ensuring consistent error responses and better debugging information.

+++ schema.sql
@@ -76,9 +76,10 @@ CREATE INDEX idx_model_versions_created_at ON model_versions(created_at DESC);
 CREATE UNIQUE INDEX idx_model_versions_active_unique ON model_versions(is_active) WHERE is_active = TRUE;
 
 -- Bot state table: tracks which accounts have active trading bots (Heartbeat architecture)
+-- account_id is BIGINT to match accounts.id (Neon/Vercel often use BIGSERIAL)
 CREATE TABLE IF NOT EXISTS bot_state (
     id SERIAL PRIMARY KEY,
-    account_id INTEGER NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
+    account_id BIGINT NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
     is_running BOOLEAN DEFAULT FALSE,
     always_on BOOLEAN DEFAULT FALSE,
     last_heartbeat TIMESTAMPTZ,
